move "translation_functions","translation_dicts" to CORE
