+++ mmd_scripting/core/translation_dictionaries.py
